Update `@dispatch` documentation
`@dispatch` processes the same way as any other event, with the current state passed as the first variable.

This lead to a bit of confusion when I attempted to use this library for the first time, until I realized that state was being passed as well. This may or may not be the desired behavior, but I'm updating the documentation to how it has worked since implementation.